added optimisations to TransformationDatabase

+++ src/darjeeling/transformation.py
@@ -1,8 +1,8 @@
-from typing import List
+from typing import List, Iterator, Dict
 from bugzoo.bug import Bug
 from bugzoo.coverage import FileLine
-from darjeeling.donor import DonorPool
 from darjeeling.donor import Snippet, DonorPool
+from darjeeling.source import SourceFile
 
 
 class Transformation(object):
@@ -19,6 +19,9 @@ class DeleteTransformation(object):
     def line(self) -> FileLine:
         return self.__line
 
+    def __str__(self) -> str:
+        return "DELETE[{}]".format(self.__line)
+
 
 class ReplaceTransformation(object):
     """
@@ -27,6 +30,9 @@ class ReplaceTransformation(object):
     def __init__(self, line: FileLine, snippet: Snippet) -> None:
         self.__line = line
         self.__snippet = snippet
+
+    def __str__(self) -> str:
+        return "REPLACE[{}; {}]".format(self.__line, self.__snippet)
 
 
 class AppendTransformation(object):
@@ -37,6 +43,9 @@ class AppendTransformation(object):
         self.__line = line
         self.__snippet = snippet
 
+    def __str__(self) -> str:
+        return "APPEND[{}; {}]".format(self.__line, self.__snippet)
+
 
 class TransformationDatabase(object):
     """
@@ -45,17 +54,44 @@ class TransformationDatabase(object):
     @staticmethod
     def generate(bug: Bug,
                  pool: DonorPool,
+                 sources: Dict[str, SourceFile],
                  lines: List[FileLine]) -> None:
         transformations = []
         for line in lines:
+            # get the content of the line
+            original_line = sources[line.filename][line.num].strip()
+
+            # ignore any line that might not be a statement
+            if not original_line.endswith(';'):
+                continue
+
+            # ignore comments and macros (probably redundant)
+            if original_line.startswith('#') or \
+               original_line.startswith('//') or \
+               original_line.startswith('/*') \
+            :
+               continue
+
             # deletion
             transformations.append(DeleteTransformation(line))
 
             # replace and append
             for snippet in pool:
-                transformations.append(ReplaceTransformation(line, snippet))
                 transformations.append(AppendTransformation(line, snippet))
+
+                # don't replace a line with an equivalent one
+                if snippet.content != original_line:
+                    transformations.append(ReplaceTransformation(line, snippet))
+
         return TransformationDatabase(transformations)
 
     def __init__(self, transformations: List[Transformation]):
         self.__transformations = transformations[:]
+
+    def __iter__(self) -> Iterator[Transformation]:
+        """
+        Returns an iterator over the transformations contained within this
+        database.
+        """
+        for t in self.__transformations:
+            yield t
